python/src/requirements.txt
mysql-connector
boto3
pandas
trycourier